requirements.txt
streamlit
matplotlib
biopython
